Add files separated out of multi-unit files.
