+++ src/main/scala/geotrellis/process/TestServer.scala
@@ -1,0 +1,6 @@
+package geotrellis.process
+
+object TestServer {
+  def apply() = Server("test", Catalog.empty("test"))
+  def apply(path:String) = new Server("test", Catalog.fromPath(path))
+}
