+++ src/main/scala/geotrellis/raster/op/tiles/ThoroughputLimitedReducer1.scala
@@ -8,24 +8,6 @@ import geotrellis.statistics._
 import geotrellis.raster._
 import geotrellis.feature.Polygon
 
-abstract class Reducer1[B: Manifest, C: Manifest](r: Op[Raster])(handle: Raster => B)(reducer: List[B] => C) extends Op[C] {
-  def _run(context: Context) = runAsync('init :: r :: Nil)
-
-  val nextSteps: Steps = {
-    case 'init :: (r: Raster) :: Nil => init(r)
-    case 'reduce :: (bs:  List[_]) => Result(reducer(bs.asInstanceOf[List[B]]))
-  }
-
-  def init(r: Raster) = {
-    r.data match {
-      case _: TiledRasterData => runAsync('reduce :: r.getTileList.map(mapper))
-      case _ => Result(reducer(handle(r) :: Nil))
-    }
-  }
-
-  def mapper(r: Raster): Op[B] = logic.Do1(r)(handle)
-}
-
 trait ThroughputLimitedReducer1[C] extends Op[C] {
   type B
   
@@ -81,27 +63,6 @@ trait ThroughputLimitedReducer1[C] extends Op[C] {
   }
 }
 
-case class TileMin(r: Op[Raster]) extends Reducer1(r)({
-  r =>
-    {
-      var zmin = Int.MaxValue
-      r.foreach {
-        z => if (z != NODATA) zmin = min(z, zmin)
-      }
-      zmin
-    }
-})({
-  zs => zs.reduceLeft((x, y) => min(x, y))
-})
+/*
 
-case class TileMax(r: Op[Raster]) extends Reducer1(r)({
-  r => r.findMinMax._2
-})({
-  zs => zs.reduceLeft((x, y) => max(x, y))
-})
-
-case class TileHistogram(r: Op[Raster]) extends Reducer1(r)({
-  r => FastMapHistogram.fromRaster(r)
-})({
-  hs => FastMapHistogram.fromHistograms(hs)
-})
+*/
